+++ native/packages/react-native-bpk-component-navigation-bar/src/BpkNavigationBar-test.ios.js
@@ -44,6 +44,25 @@ describe('ios', () => {
       expect(tree).toMatchSnapshot();
     });
 
+    it('should render correctly with custom style', () => {
+      const tree = renderer
+        .create(
+          <BpkNavigationBar
+            title="Backpack"
+            leadingButton={
+              <BpkNavigationBarBackButtonIOS
+                title="Back"
+                showTitle
+                onPress={jest.fn()}
+              />
+            }
+            style={{ zIndex: 5 }}
+          />,
+        )
+        .toJSON();
+      expect(tree).toMatchSnapshot();
+    });
+
     it('should render correctly with trailing button', () => {
       const tree = renderer
         .create(